Fix compile errors from TV example on Ubuntu x86-64 (#6914)
* Fix compile errors from TV example on Ubuntu x86-64

* Update examples/tv-app/linux/include/tv-channel/TvChannelManager.cpp


* Update examples/tv-app/linux/include/target-navigator/TargetNavigatorManager.cpp


* Update examples/tv-app/linux/include/media-input/MediaInputManager.cpp


* Update examples/tv-app/linux/include/audio-output/AudioOutputManager.cpp


* Use int instead of uint_8/uint_16 as loop index

+++ examples/tv-app/linux/include/audio-output/AudioOutputManager.cpp
@@ -52,12 +52,12 @@ vector<EmberAfAudioOutputInfo> AudioOutputManager::proxyGetListOfAudioOutputInfo
     int maximumVectorSize = 3;
     char name[]           = "exampleName";
 
-    for (uint8_t i = 0; i < maximumVectorSize; ++i)
+    for (int i = 0; i < maximumVectorSize; ++i)
     {
         EmberAfAudioOutputInfo audioOutputInfo;
         audioOutputInfo.outputType = EMBER_ZCL_AUDIO_OUTPUT_TYPE_HDMI;
         audioOutputInfo.name       = chip::ByteSpan(chip::Uint8::from_char(name), sizeof(name));
-        audioOutputInfo.index      = 1 + i;
+        audioOutputInfo.index      = static_cast<uint8_t>(1 + i);
         audioOutputInfos.push_back(audioOutputInfo);
     }
